Almost finished. Still need to update Reports, if that table is still being used (last entry for me is 07/2015)
Added a model for the join table used in the Documents,
SubServiceRequests has_and_belongs_to_many relationship, so copying
Documents to new SubServiceRequests can be audited. Added Audit
comments for the big changes that are made in the rake task; Wenjun or
Leila want to keep a specific historical record. Also, there are no
Subsidies for the SSR’s that are being split, and ResponseSets do not
appear to belong to any SSR, despite having a sub_service_request_id
column.